Remove bootstrap.sh version pinning.
This utility is intended to be run on fresh installs, with potentially new default
apt repositories - this is at least the case with the Docker local test framework.

It'd be great to mandate versioning _better_, but it feels like this needs more
thought than can be put it in at this stage.