fix(spotify): Update signing key
Updates the signing key the current version, as well as shifts to a
trusted key approach rather than signed by, per the Spotify installation
docs.

We should also be dearmoring the asc's but eh, that's much more work for
no net gain at the moment - we at least have the gpg package now, so
it is feasible.