refac: standard role inclusion for package metadata
This commit rewrites the typical pattern used for establishing which
packages are already installed, and includes as part of the apt role.

To support this, the apt role tag logic was rejigged to prevent tasks
being included accidentally.

This refactoring simplifies things significantly, and results in a
greater degree of yamllint compliance.