refac: removed GNOME configuration
While the idea of managing taskbar favorites via a script *is* cool,
this implementation is not fit for purpose.

To do this properly, we'd want to write something in Python - either a
standalone script, or an action plugin.

Until then, though, this code is just technical debt, so we may as well
remove it.