refactor: use documented source to simplify apt
By exploiting the documented default paths for upstream debian sources,
we can avoid a lot of the find and replace style work that we were
doing.

This works by diffing out managed templated against the deployed source
(in the documented location) and dpeloying our iteration if changes
exist.

This significantly reduces the scope of the role.

Docs:
* https://wiki.debian.org/SourcesList#debian.sources