feat: provide 'remote' dev environment
This commit makes some questionable decisions and defines a new Docker
Compose service with two containers as a parity testing environment for
working on remote deployments.

The main 'magic' is mounting SSH keys into the container - in the
hindsight a better approach would be to mount the users *actual* ssh
keys - then we don't need to manage anything in the compose spec.

Anyway - it works; it provides an env for which we can run the
remote.yaml playbook without bricking our actual remote VMs